requirements.txt
streamlit
supabase
qrcode
pillow
fpdf2
pytz
pandas
